refactor: 수락과 거절을 다른 메서드로 처리하도록 수정

+++ backend/src/main/java/backend/mulkkam/friend/service/FriendService.java
@@ -8,7 +8,6 @@ import backend.mulkkam.common.dto.MemberDetails;
 import backend.mulkkam.common.exception.CommonException;
 import backend.mulkkam.friend.domain.Friend;
 import backend.mulkkam.friend.domain.FriendRequest;
-import backend.mulkkam.friend.domain.RequestDecision;
 import backend.mulkkam.friend.repository.FriendRepository;
 import backend.mulkkam.friend.repository.FriendRequestRepository;
 import backend.mulkkam.member.repository.MemberRepository;
@@ -25,9 +24,8 @@ public class FriendService {
     private final MemberRepository memberRepository;
 
     @Transactional
-    public void processFriendRequest(
+    public void acceptFriendRequest(
             Long friendRequestId,
-            RequestDecision requestDecision,
             MemberDetails memberDetails
     ) {
         Long memberId = memberDetails.id();
@@ -38,10 +36,24 @@ public class FriendService {
             throw new CommonException(NOT_PERMITTED_FOR_PROCESS_FRIEND_REQUEST);
         }
 
-        if (requestDecision.equals(RequestDecision.ACCEPT)) {
-            Friend friend = new Friend(friendRequest);
-            friendRepository.save(friend);
+        Friend friend = new Friend(friendRequest);
+        friendRepository.save(friend);
+        friendRequestRepository.delete(friendRequest);
+    }
+
+    @Transactional
+    public void rejectFriendRequest(
+            Long friendRequestId,
+            MemberDetails memberDetails
+    ) {
+        Long memberId = memberDetails.id();
+        validateMemberExistence(memberId);
+        FriendRequest friendRequest = getFriendRequest(friendRequestId);
+
+        if (!friendRequest.validatePermission(memberId)) {
+            throw new CommonException(NOT_PERMITTED_FOR_PROCESS_FRIEND_REQUEST);
         }
+
         friendRequestRepository.delete(friendRequest);
     }
